Update MK8 Cluster Fuel Resistor Requirements
The current instructions request a 180ohm resistor connection to display full fuel. However an 180ohm resistor displays empty on the cluster.

This is an update to the readme to clarify that a 10ohm resistor is required for full fuel. This was identified through digging through the xconnect discord and functional testing